COSBETA-24: Use analyzer only, no job
